Remove bad_address logic
The bad_address signal was being used to mask out the enable signal when a switch pattern not mapping to a channel was detected. This caused continuous conversion to stop, and there was no method of restarting it without reprogramming the demo into the board. This commit removes bad_address and relies on the default case to ensure that a valid address is provided. This was a faster fix than the alternative of implementing a mechanism to restart once a valid switch position was found again.